Connection adapter #tables deprecation warnigns are classified
Connection adapter #tables warnings are now matched and supressed.
Extracted some common code into a base class for classifiers.

Currently the supression is being done directly in the classifier by
returning a nil deprecation warning in the classifier result - this
should be moved into a seperate concept and handled in the log analyser.