Fixed bugs and refactored classifiers
LogParser now works correctly with LogCursor.
Renamed log_lines to cursor everywhere as we are now passing the cursor
around.
Moved the common classifier operations up into the base classifier and
utilised a template approach to simplify the creation of classifiers and
reduce the amount of knowledge each classifier needed.