Simplify classifier configuration
Previously lines_to_consume and summary were required as part of the
simple classifier configuration yaml, but lines_to_consume is almost
always 1 and summary is almost always the first sentence of the message.

Update the logic that builds simple classifier instances from yaml to
insert default values for lines_to_consume and summary if they are not
present in the config.